default username can be set
By default, the user name will be parsed from the input file containing lines formatted like user:password

Using --username/-u a default username can be set which is then used during the SSH authentication process. Parsing of the input file will change in a way that each line of the file is treated as the password.
This would allow to use dictionary files such as /usr/share/dict/words for testing without the need to preprocess them.