support CLI and programmatic run mode
If script is started from the command line the arguments will be parsed automatically and the tests will be run thereafter.

If initialized programatically, neither the arguments will be passed (because they are not provided yet) nor the tests will be started automatically. You need to do it from the programm which uses the password checker.

Some old stuff was removed which has not been used.